requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
PyMySQL==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0